python-backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.23.2

# PDF Parsing and OCR
PyPDF2==3.0.1
PyMuPDF==1.23.7
pymupdf4llm==0.1.1

# Language Detection
langdetect

# Tokenization and Embeddings
tiktoken
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0

# NLP for Entity Extraction
spacy==3.7.2
nltk==3.8.1

# Database Interaction
sqlalchemy==2.0.23
pgvector==0.2.3
psycopg2-binary==2.9.9

# Logging and Utilities
Pillow
html5lib
lxml
beautifulsoup4
einops
python-dotenv==1.0.0
numpy==1.26.1
pandas==2.1.2

# CORS
starlette==0.27.0

# Authentication and Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Testing
pytest==7.4.3

# HTTP Client
httpx==0.25.1